+++ api/models/Client.go
@@ -10,7 +10,7 @@ import (
 )
 
 type Client struct {
-	ID    uint32 `gorm:"primary_key;auto_increment" json:"id"`
+	ID    uint64 `gorm:"primary_key;auto_increment" json:"id"`
 	Name  string `gorm:"size:255;not null; unique" json:"name"`
 	Email string `gorm:"size:100;not null;unique" json:"email"`
 	// Receipts  Receipt   `gorm:"foreignKey:UserRefer json:receipt"`
@@ -68,7 +68,7 @@ func (c *Client) FindClientByID(db *gorm.DB, cid uint64) (*Client, error) {
 	return c, err
 }
 
-func (c *Client) UpdateAClient(db *gorm.DB, cid uint32) (*Client, error) {
+func (c *Client) UpdateAClient(db *gorm.DB, cid uint64) (*Client, error) {
 	var err error
 	db = db.Debug().Model(&Client{}).Where("id = ?", cid).Take(&Client{}).UpdateColumn(
 		map[string]interface{}{
